+++ DetourWrapper/DllExport.h
@@ -22,7 +22,7 @@ extern "C"
     // Updated to match new signatures with include/exclude flags
     DETOUR_API uint32_t find_path(void* ptr, void* start, void* end, uint16_t includeFlags, uint16_t excludeFlags, float* strPath);
     DETOUR_API uint32_t find_smoothPath(void* ptr, void* start, void* end, uint16_t includeFlags, uint16_t excludeFlags, float* smoothPath);
-    DETOUR_API uint32_t random_point(void* ptr, void* centerPoint, float radius, uint16_t includeFlags, uint16_t excludeFlags, float* rndPoint);
+    DETOUR_API uint32_t random_point(void* ptr, void* centerPoint, int radius, uint16_t includeFlags, uint16_t excludeFlags, float* rndPoint);
     DETOUR_API uint32_t getPolyFlags(void* ptr, void* pos, uint16_t includeFlags, uint16_t excludeFlags);
 
     // LOS does not take flags — unchanged
